Refactor raw log decoding method
Existing logic required Wamore's log decoding executable to be in the
same directory as the log file to be decoded. This is a holdover from
the previously held assumption that it did not accept full file names,
which has been tested false. The current assumption is now that the
executable resides in MATLAB's working directory when executing iDAQ.m

This change also creates a clear path forward for when AppLocker is
implemenetd on DREN machines, requiring all executables to run from
a whitelisted directory.