Fix issues with descent rate curve axes limiting
The data windowing method plots the data of interest against the data
index to simplify indexing the windowed data. In the case of the descent
rate calculation method, the data is replotted against time for better
vieweing. The axes limits are then adjusted based on the limits of the
axes being used while windowing the data. Fixes have been implemented
for cases where these axes limits correspond to invalid indices, either
indices less than 1 or indices greater than the length of the data set.

Resolves #2